Revert back to 500 status code, fixing server status code
The code was altered from a 500 to a 400 status code as per my
suggestion.  Given other checks in this code, that suggestion seems
wrong.

We had to go into the code anyways because the json message was a 400
and the returned code was a 500.

While we were at it, we also added this to the method's documentation
so it's easier to grok for bystanders when checking the code itself.
The README has also been updated to contain said information.