Split delete query in two parts.
Given a file residing in multiple graphs (> 2), the original query
didn't perform well when it gets transformed by mu-auth and sent to
virtuoso.

The problem is related to virtuoso, which estimates the
cost of executing the transformed query as extraordinarily high and
therefore, blocking the execution.

I am aware the need for a re-write is implementation-specific and the
solution might not work if the amount of graphs keeps on growing.

However, virtuoso isn't going to go away soon, it is working better
now and the result - I assume - of the re-write is the logical equivalent of the
original query.

There is a risk for unexpected consequences on how
deltas are propagated through mu-auth (one block of deletes/inserts, vs
multiple blocks?) and potentially affect consuming services.
I didn't test for that, but in my defence, I would state:
  - mu-auth potentially might eventually fold the resulting changesets
of such calls
  - Services should be robust and not assume transaction-like blocks
in the handling of incoming deltas. (i.e. be in line with the general
philosophy of reactive microservices in a semantic.works stack)

This rewrite will be tested in some of the deployed stacks, so we'll
have an idea of its validity soon.